+++ example/main.dart
@@ -27,8 +27,23 @@ class TextFieldDemoPage extends StatefulWidget {
 class _TextFieldDemoPageState extends State<TextFieldDemoPage> {
   final TextEditingController _usernameController = TextEditingController();
   final TextEditingController _passwordController = TextEditingController();
+  final TextEditingController _emailController = TextEditingController();
   final FocusNode _usernameFocus = FocusNode();
   final FocusNode _passwordFocus = FocusNode();
+  final FocusNode _emailFocus = FocusNode();
+
+  MobileNumberInfo? _phoneNumber;
+
+  @override
+  void dispose() {
+    _usernameController.dispose();
+    _passwordController.dispose();
+    _emailController.dispose();
+    _usernameFocus.dispose();
+    _passwordFocus.dispose();
+    _emailFocus.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -39,19 +54,60 @@ class _TextFieldDemoPageState extends State<TextFieldDemoPage> {
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            TextFormField(
-              controller: _usernameController,
+            AppTextInput(
+              labelText: 'Username',
+              hintText: 'Enter your username',
+              textController: _usernameController,
               focusNode: _usernameFocus,
-              textInputAction: TextInputAction.next,
-              decoration: InputDecoration(
-                labelText: 'Username',
-                hintText: 'Enter your username',
-              ),
-              onFieldSubmitted: (_) => InputFocusHandler.changeFocusField(
-                context,
-                _usernameFocus,
-                _passwordFocus,
-              ),
+              nextFocus: _emailFocus,
+              onSubmitted:
+                  (value) => InputFocusHandler.changeFocusField(
+                    context,
+                    _usernameFocus,
+                    _emailFocus,
+                  ),
+            ),
+            const SizedBox(height: 20),
+            AppTextFormField(
+              labelText: 'Email',
+              hintText: 'Enter your email',
+              keyboardType: TextInputType.emailAddress,
+              textController: _emailController,
+              focusNode: _emailFocus,
+              nextFocus: _passwordFocus,
+              validator:
+                  (value) =>
+                      (value == null || !value.contains('@'))
+                          ? 'Invalid email'
+                          : null,
+              onChanged: (value) {},
+              onFieldSubmitted:
+                  () => InputFocusHandler.changeFocusField(
+                    context,
+                    _emailFocus,
+                    _passwordFocus,
+                  ),
+            ),
+            const SizedBox(height: 20),
+            AppPhoneField(
+              initialCountryCode: 'US',
+              onChanged: (MobileNumberInfo? info) {
+                setState(() {
+                  _phoneNumber = info;
+                });
+
+                debugPrint(
+                  'phoneNumber: ${_phoneNumber?.completeNumber} info: ${info.toString()}',
+                );
+              },
+              validator: (value) {
+                if (value == null || value.isEmpty)
+                  return 'Phone number required';
+                return null;
+              },
+              focusNode: FocusNode(),
+              nextFocus: _passwordFocus,
+              errorMsg: null,
             ),
             const SizedBox(height: 20),
             PasswordField(
@@ -60,6 +116,7 @@ class _TextFieldDemoPageState extends State<TextFieldDemoPage> {
               labelText: 'Password',
               hintText: 'Enter your password',
               validator: (value) => value!.isEmpty ? 'Password required' : null,
+              nextFocus: null,
             ),
             const SizedBox(height: 20),
             PinCodeField(